+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v0.8.2 - DDMMM24
+# v0.8.2 - 10DEC24
 
 ## Added Features